Move getChannel to each test
